+++ apps/console/src/protected-app.tsx
@@ -199,6 +199,7 @@ export const ProtectedApp: FunctionComponent<AppPropsInterface> = (): ReactEleme
             let response: BasicUserInfo = { ...signInResponse };
             let logoutUrl: string;
             let logoutRedirectUrl: string;
+            let subOrgIdToken: string;
 
             const idToken: DecodedIDTokenPayload = await getDecodedIDToken();
             const isPrivilegedUser: boolean =
@@ -302,6 +303,7 @@ export const ProtectedApp: FunctionComponent<AppPropsInterface> = (): ReactEleme
                     );
                     
                     setTenant(response.orgId);
+                    subOrgIdToken = response.orgId;
                     dispatch(setCurrentOrganization(response.orgName));
                 }
             }
@@ -437,7 +439,7 @@ export const ProtectedApp: FunctionComponent<AppPropsInterface> = (): ReactEleme
             dispatch(
                 setSignIn<AuthenticatedUserInfo & TenantListInterface>(
                     Object.assign(
-                        CommonAuthenticateUtils.getSignInState(response, response.orgId),
+                        CommonAuthenticateUtils.getSignInState(response, subOrgIdToken),
                         {
                             associatedTenants: isPrivilegedUser
                                 ? tenantDomain